requirements_structural.txt
# STRUCTURAL FACE ANALYSIS REQUIREMENTS
# Install with: pip install -r requirements_structural.txt

# Core dependencies
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0

# Optional (for web integration)
flask==3.0.0

# Optional (for advanced features)
matplotlib==3.8.2
